cleaned up contributor routes
